+++ NEWS.md
@@ -1,5 +1,7 @@
 # cowplot 1.2.0
 
+- fix `get_legend()` for redesigned ggplot legend system (#202)
+
 - fix test for ggplot object class (#210)
 
 - fix label order when laying out plots by column rather than by row